Extract state machine for classifying TS files
